refactor(tests): Consolidate redundant mapState tests into proper container tests
Reduces test duplication by removing 6 *.mapState.test.js files that tested
the same getVisibleBoards() filtering behavior already covered in Board.selectors.test.js.

Changes:
- Created Board.container.test.js with unique test for active board returning undefined when soft-deleted
- Created CommunicatorDialog.container.test.js with unique test for communicatorBoards two-step filtering
- Added documentation comments in container files explaining test organization
- Deleted 6 redundant test files (Analytics, Board, CommunicatorDialog, CommunicatorToolbar, Export, Import)

Benefits:
- Reduces maintenance burden (7 files → 3 files)
- Eliminates 306 lines of redundant test code
- Improves test organization following standard patterns (selectors → *.selectors.test.js, containers → *.container.test.js)
- Makes it clear what's tested where through inline comments

Test coverage maintained: All unique tests preserved, redundant tests removed.
