+++ src/components/Board/__tests__/Board.container.test.js
@@ -1,0 +1,72 @@
+import { mapStateToProps } from '../Board.container';
+import { SYNC_STATUS } from '../Board.constants';
+
+jest.mock('ogv', () => ({ OGVLoader: { base: '' } }));
+jest.mock('dom-to-image', () => ({}));
+jest.mock('../Board.component', () => () => null);
+
+const createState = (boards, syncMeta = {}) => ({
+  board: {
+    boards,
+    syncMeta,
+    activeBoardId: null,
+    output: [],
+    navHistory: [],
+    isLiveMode: false,
+    improvedPhrase: null
+  },
+  communicator: {
+    activeCommunicatorId: 'comm-1',
+    communicators: [{ id: 'comm-1', boards: boards.map(b => b.id) }]
+  },
+  speech: {
+    voices: [],
+    options: { voiceURI: null, isCloud: false }
+  },
+  scanner: {},
+  app: {
+    displaySettings: {},
+    navigationSettings: {},
+    userData: null,
+    isConnected: true,
+    liveHelp: {
+      isRootBoardTourEnabled: false,
+      isSymbolSearchTourEnabled: false,
+      isUnlockedTourEnabled: false
+    }
+  },
+  language: { lang: 'en-US' },
+  subscription: {
+    premiumRequiredModalState: null,
+    isInFreeCountry: true,
+    isSubscribed: false,
+    isOnTrialPeriod: false
+  }
+});
+
+describe('Board.container', () => {
+  describe('mapStateToProps', () => {
+    describe('active board handling', () => {
+      it('returns undefined for active board that is soft-deleted', () => {
+        const boards = [{ id: 'board-1' }, { id: 'board-2' }];
+        const syncMeta = {
+          'board-2': { status: SYNC_STATUS.PENDING, isDeleted: true }
+        };
+        const state = {
+          ...createState(boards, syncMeta),
+          board: {
+            ...createState(boards, syncMeta).board,
+            activeBoardId: 'board-2'
+          }
+        };
+
+        const props = mapStateToProps(state);
+
+        expect(props.board).toBeUndefined();
+      });
+
+      // Note: Basic getVisibleBoards() filtering logic is tested in Board.selectors.test.js
+      // This test covers the container-specific behavior: board: getVisibleBoards(state).find(board => board.id === activeBoardId)
+    });
+  });
+});
