--- src/components/Communicator/CommunicatorToolbar/CommunicatorToolbar.container.mapState.test.js
@@ -1,44 +1,0 @@
-import { mapStateToProps } from './CommunicatorToolbar.container';
-import { SYNC_STATUS } from '../../Board/Board.constants';
-
-const createState = (boards, syncMeta = {}) => ({
-  board: {
-    boards,
-    syncMeta,
-    activeBoardId: null
-  },
-  communicator: {
-    activeCommunicatorId: 'comm-1',
-    communicators: [{ id: 'comm-1', boards: boards.map(b => b.id) }]
-  },
-  app: {
-    userData: null,
-    displaySettings: { darkThemeActive: false }
-  }
-});
-
-describe('CommunicatorToolbar.container mapStateToProps', () => {
-  describe('boards prop', () => {
-    it('returns all boards when none are soft-deleted', () => {
-      const boards = [{ id: 'board-1' }, { id: 'board-2' }];
-      const state = createState(boards);
-
-      const props = mapStateToProps(state, {});
-
-      expect(props.boards).toHaveLength(2);
-    });
-
-    it('excludes soft-deleted boards', () => {
-      const boards = [{ id: 'board-1' }, { id: 'board-2' }];
-      const syncMeta = {
-        'board-2': { status: SYNC_STATUS.PENDING, isDeleted: true }
-      };
-      const state = createState(boards, syncMeta);
-
-      const props = mapStateToProps(state, {});
-
-      expect(props.boards).toHaveLength(1);
-      expect(props.boards[0].id).toBe('board-1');
-    });
-  });
-});
